### 2021-09-03 #3683 @whedon
[PRE REVIEW]: BioPsyKit: A Python package for the analysis of biopsychological data
**Submitting author:** @richrobe (<a href="http://orcid.org/0000-0003-0272-5403">Robert Richer</a>)
**Repository:** <a href="https://github.com/mad-lab-fau/BioPsyKit" target ="_blank">https://github.com/mad-lab-fau/BioPsyKit</a>
**Version:** 0.3.0
