### 2021-09-08 #3683 @espenhgn
> @espenhgn @torbjone would any of you be interested in reviewing this submission to the Journal of Open Source Software? It seems to be within your areas of expertise.

Hi @osorensen, I think I can do this if you need another reviewer.
