### 2021-09-08 #3702 @whedon
[REVIEW]: BioPsyKit: A Python package for the analysis of biopsychological data
**Submitting author:** @richrobe (<a href="http://orcid.org/0000-0003-0272-5403">Robert Richer</a>)
**Repository:** <a href="https://github.com/mad-lab-fau/BioPsyKit" target ="_blank">https://github.com/mad-lab-fau/BioPsyKit</a>
**Version:** 0.3.2
**Archive:** <a href="https://doi.org/10.17605/OSF.IO/EBDRT" target="_blank">10.17605/OSF.IO/EBDRT</a>

- [x] #3748
- [x] #3749

- [x] #3750
- [x] #3751
