### 2021-09-13 #3702 @zen-juen
Hi there! First I'd like to declare that I am one of the core team members of another software package, [NeuroKit2](https://github.com/neuropsychology/NeuroKit), of which @richrobe has made some contributions to, and some of neurokit's functions are also subsumed under BioPsyKit as mentioned in the manuscript. That said, I will try to remain fully objective in my review here. @osorensen am I able to continue with the review?
