### 2021-09-13 mad-lab-fau/BioPsyKit#10 @zen-juen
[JOSS review] Get example datasets in README error
Hi there, after installing `biopsykit` with ease, I've encountered some errors when running the examples listed on the readme, particularly those that are related to retrieving example datasets:

```
import biopsykit as bp
from biopsykit.example_data import get_saliva_example

saliva_data = get_saliva_example(sample_times=[-20, 0, 10, 20, 30, 40, 50])
```
I get an error like so:
```

  File "c:\users\zen juen\downloads\wpy64-3850\python-3.8.5.amd64\lib\site-packages\pandas\io\parsers.py", line 1862, in __init__
    self._open_handles(src, kwds)

  File "c:\users\zen juen\downloads\wpy64-3850\python-3.8.5.amd64\lib\site-packages\pandas\io\parsers.py", line 1357, in _open_handles
    self.handles = get_handle(

  File "c:\users\zen juen\downloads\wpy64-3850\python-3.8.5.amd64\lib\site-packages\pandas\io\common.py", line 642, in get_handle
    handle = open(

FileNotFoundError: [Errno 2] No such file or directory: 'c:\\users\\zen juen\\downloads\\wpy64-3850\\python-3.8.5.amd64\\lib\\example_data\\cortisol_sample.csv'
```
This also happens for `get_sleep_imu_example()`, `get_questionnaire_example()` and `get_mist_hr_example()`. Might this be to do with the way functions are setting the data path?

(Note: as part of https://github.com/openjournals/joss-reviews/issues/3702)
