### 2021-09-13 mad-lab-fau/BioPsyKit#11 @zen-juen
[JOSS REVIEW] Documentation
Hi there, when ticking off the documentation checks on https://github.com/openjournals/joss-reviews/issues/3702 I realize that there aren’t guidelines yet for potential contributors, collaborators, and those seeking support for software use. You may want to refer to https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors for doing so, as well as add some templates for writing issues and creating pull requests!

Overall, I think the README is clear with short and succinct examples of use and are easily referred to when going through each section on different types of analysis in the paper (e.g., sleep analysis, biomarker analysis…). However, there is no documentation of how statistical analyses and machine learning pipelines (using `stats` and `classification`) can be readily implemented. As it's not intuitive yet on how this can be done, I think a short example script could be nice - perhaps the one used to generate the Figure 5 boxplot in the paper?

Let me know what you think!
