### 2021-09-22 mad-lab-fau/BioPsyKit#14 @espenhgn
[JOSS Review] Broken example notebooks
There are several errors and issues with the provided example notebooks:
- There are many abbreviations like IMU, PSS, PANAS, HRV that are not explained. Perhaps these are standard terms in the field, but for a broader acceptance of this tool, these abbreviations should be defined.
- `ECG_Processing_Example.ipynb`: throws an error at the line `fig, axs = ecg.plotting.ecg_plot(ep, key='Data', figsize=(10,5))`; `NameError: name 'ecg' is not defined`
- `ECG_Analysis_Example.ipynb` fails on line `display_dict_structure(study_data_dict)` because of the above error in `ECG_Processing_Example.ipynb` (no results to load).
- `EEG_Example.ipynb`: Please add some documentation of what functionality this notebook is supposed to demonstrate. Some time series are shown but it’s not very clear what is going on.
- `Log_Data_Example.ipynb`; ditto.
- `Protocol_Example.ipynb` fails because of the above error in `ECG_Processing_Example.ipynb` (no results to load)
- `Questionnaire_Example.ipynb`: Incorporating purposely failing code (`pss = bp.questionnaires.pss(data_pss)` resulting in `ValueRangeError …` ) in the notebook, I don’t think that’s a good idea whatsoever (prevents running all cells and such). One could capture the error in a more controlled manner.
- `SklearnPipelinePermuter_Example.ipynb`: Throws an error `NameError: name 'datasets' is not defined` on line `breast_cancer = datasets.load_breast_cancer()`
- `StatsPipeline_Plotting_Example.ipynb`: Typo “cortsol”
- `STROOP_Example.ipynb`: Crashes at line `stroop.hr_ensemble_plot(data=dict_phase,figsize=(10,8),ylims=(40,120))` with `TypeError: hr_ensemble_plot() missing 1 required positional argument: 'ensemble_id'`

Linked issue: https://github.com/openjournals/joss-reviews/issues/3702
