### 2021-09-22 #3702 @kthyng
Hi @espenhgn! I am the managing editor on rotation this week and it looks like you're having some trouble with the mechanics of the review here. Unfortunately the little "convert to issue" buttons that github provides aren't connected with the JOSS workflow, which is why the issues are immediately closed (since we can't turn off the buttons). Also, is [this issue](https://github.com/openjournals/joss-reviews/issues/3751) meant to be in regard to your review? If so, please open it instead in the software repo, https://github.com/mad-lab-fau/BioPsyKit, and close it where it is currently. Do link it back to this review issue though. Please let me or @osorensen know if you have any questions about this process.
