### 2021-09-23 #3702 @zen-juen
Hi all!

Firstly, I really like that `BioPsyKit` nicely incorporates biophysiological and self-report methodologies in one single package, allowing for the systematic processing and analysis of multimodal data. Below are some points that will help me to check off the remaining boxes:

- The README and docstrings are well-documented with concise examples of use, installation instructions, and the package also comprises a comprehensive test suite. However, the documentation will further benefit from having clear guidelines for potential contributors/new users and an additional example on how to set up statistical analysis pipelines as it is not intuitive how this can be done yet. See https://github.com/mad-lab-fau/BioPsyKit/issues/11. Additionally, I realized the `readthedocs` page does not render (or not available yet?) despite having a `.readthedocs.yml` file.

- The paper provides a good overview of the package, but requires significantly more elaboration on how `BioPsyKit` addresses specific existing needs in biopsychology (i.e., what is the problem it is solving) and some clarification on the usage of certain submodules. I've detailed and explained what this entails here https://github.com/mad-lab-fau/BioPsyKit/issues/12

- Finally, I can't assess the functionality and performance of the software yet due to https://github.com/mad-lab-fau/BioPsyKit/issues/10 - once this is addressed, I'll be able to complete the review! 😄
