+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -1,0 +1,32 @@
+---
+name: Bug report
+about: Create a report to help us improve
+title: ''
+labels: bug
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots and Minimal Working Example**
+If applicable, add screenshots and a minimal working example (e.g., test data that caused your issue) to help explain your problem.
+
+**System (please complete the following information):**
+ - Operating System: [e.g. Windows, macOS]
+ - BioPsyKit version
+ - Python version
+
+**Additional context**
+Add any other context about the problem here.
