+++ .github/ISSUE_TEMPLATE/bug_report.md
@@ -23,10 +23,19 @@ A clear and concise description of what you expected to happen.
 **Screenshots and Minimal Working Example**
 If applicable, add screenshots and a minimal working example (e.g., test data that caused your issue) to help explain your problem.
 
-**System (please complete the following information):**
+**System Specifications (please complete the following information):**
  - Operating System: [e.g. Windows, macOS]
  - BioPsyKit version
  - Python version
+
+The easiest way to provide us these information is by running this code in a Python console:
+
+```python
+import biopsykit as bp
+bp.version()
+```
+
+and copy/paste the output.
 
 **Additional context**
 Add any other context about the problem here.
