+++ docs/source/development/development_guide.md
@@ -1,361 +1,0 @@
-# Development Guide
-
-This document contains information for developers that need further in-depth information on how to setup and use tools
-and learn about programing methods used in development of this project.
-
-[comment]: <> (If you are looking for a higher level overview over the guiding ideas and structure of this project, please visit the)
-[comment]: <> ([Project Structure document]&#40;project_structure.md&#41;.)
-
-## Project Setup and Poetry
-
-*biopsykit* only supports Python 3.7 and newer.
-First, install a compatible version of Python.
-If you do not want to modify your system installation of Python you can use [conda](https://docs.conda.io/en/latest/)
-or [pyenv](https://github.com/pyenv/pyenv).
-However, there are some issues with using conda.
-Please, check the [trouble shooting guide](#trouble-shooting) below.
-
-*biopsykit* uses [poetry](https://python-poetry.org) to manage its dependencies.
-Once you installed poetry, run the following commands to initialize a virtual env and install all development
-dependencies:
-
-```bash
-poetry install
-```
-This will create a new folder called `.venv` inside your project dir.
-It contains the python interpreter and all site packages.
-You can point your IDE to this folder to use this version of Python.
-For PyCharm you can find information about this 
-[here](https://www.jetbrains.com/help/pycharm/configuring-python-interpreter.html).
-
-**In case you encounter any issues (with this command or any command below), please check the section on
- [trouble shooting](#trouble-shooting)**.
- 
-To add new dependencies:
-
-```bash
-poetry add <package name>
-
-# Or in case of a dev dependency
-poetry add --dev <package name>
-```
-
-For more commands see the [official documentation](https://python-poetry.org/docs/cli/).
-
-To update dependencies after the `pyproject.toml` file was changed (It is a good idea to run this after a `git pull`):
-```bash
-poetry install --no-root
-
-# or (see differences below)
-poetry update
-```
-
-Running `poetry install` will only install packages that are not yet installed. `poetry update` will also check, if 
-newer versions of already installed packages exist.
-
-## Tools we are using
-
-To make it easier to run commandline tasks we use [doit](https://pydoit.org/contents.html) to provide a cross-platform 
-cli for common tasks.
-All commands need to be executed in the `venv` created by poetry.
-
-To list the available tasks, run:
-
-```bash
-$ poetry run doit list
-docs                 Build the html docs using Sphinx.
-format               Reformat all files using black.
-format_check         Check, but not change, formatting using black.
-lint                 Lint all files with Prospector.
-test                 Run Pytest with coverage.
-update_version       Bump the version in pyproject.toml and biopsykit.__init__ .
-```
-
-To run one of the commands execute (e.g. the `test` command):
-```bash
-poetry run doit test
-```
-
-To execute `format`, `lint`, and `test` all together, run:
-```bash
-poetry run doit
-# or if you want less output
-petry run doit -v 0
-```
-
-You should run this as often as possible!
-At least once before any `git push`.
-
-**Protip**: If you do not want to type `poetry run` all the time, you can also activate the `venv` for your current
-terminal session using `poetry shell`.
-After this you can just type, for example, `doit test`.
-
-To ensure that the whole library uses a consistent **format**, we use [black](https://github.com/psf/black) to
-autoformat our code.
-Black can also be integrated [into your editor](https://black.readthedocs.io/en/stable/integrations/editors.html), 
-if you do not want to run it from the command line.
-Because, it is so easy, we also use *black* to format the test-suite.
-
-For everything *black* can not handle, we use [prospector](http://prospector.landscape.io/en/master/) to handle all 
[…]
