### 2021-09-28 ef4d858
changed github workflow order: Linting -> Testing -> Coverage Badge instead of Testing -> Coverage Badge -> Linting
