+++ .github/workflows/test-and-lint.yml
@@ -26,6 +26,9 @@ jobs:
         pip install poetry
         poetry config virtualenvs.in-project true
         poetry install -E mne -E jupyter
+    - name: Linting
+      run: |
+        poetry run doit -v 0 --continue format_check lint
     - name: Testing
       run: |
         poetry run doit test
@@ -46,7 +49,3 @@ jobs:
       with:
         github_token: ${{ secrets.GITHUB_TOKEN }}
         branch: ${{ github.ref }}
-    - name: Linting
-      run: |
-        poetry run doit -v 0 --continue format_check lint
-
