### 2021-09-28 e380e22
dependencies: changed minimal required scipy version from ^1.7.1 to ^1.7.0
