+++ pyproject.toml
@@ -24,7 +24,7 @@ matplotlib = "^3"
 seaborn = "^0.11.1"
 neurokit2 = "^0.1.3"
 pingouin = "^0.4.0"
-scipy = "^1.7.1"
+scipy = "^1.7.0"
 statannot = "^0.2.3"
 tqdm = "^4.62.0"
 joblib = "^1.0.0"
