+++ .github/PULL_REQUEST_TEMPLATE/pull_request_template.md
@@ -1,0 +1,44 @@
+This is a template for creating a Pull Request. You can remove the text and sections if you wish, just make sure you give enough information about the how and why of the Pull Request.
+
+
+# Description
+
+Please include a summary of the change and which issue is fixed. Please also include relevant motivation and context. List any dependencies that are required for this change.
+
+## Type of change
+
+I changed the `foo()` function so that ...
+
+or
+
+Fixes # (issue)
+
+Please delete options that are not relevant.
+
+- [ ] Bug fix (non-breaking change which fixes an issue)
+- [ ] New feature (non-breaking change which adds functionality)
+- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
+- [ ] This change requires a documentation update
+
+# Checklist:
+
+- [ ] My code follows the style guidelines of this project
+- [ ] I have performed a self-review of my own code
+- [ ] I have commented my code, particularly in hard-to-understand areas
+- [ ] I have made corresponding changes to the documentation
+- [ ] My changes generate no new warnings
+- [ ] I have added tests that prove my fix is effective or that my feature works
+- [ ] New and existing unit tests pass locally with my changes
+- [ ] Any dependent changes have been merged and published in downstream modules
+
+
+
+
+# Checklist
+
+Here are some things to check before creating the PR. If you encounter any issues, do let us know :)
+
+- [ ] I have read the [CONTRIBUTING](https://github.com/mad-lab-fau/BioPsyKit/blob/main/CONTRIBUTING.md) file.
+- [ ] My Pull Request is targetted at the **`main` branch**.
+- [ ] My Pull Request passes all format, linting and test checks.
+- [ ] I have added the newly added features to the [CHANGELOG](https://github.com/mad-lab-fau/BioPsyKit/blob/main/CHANGELOG.md) (if applicable)
