+++ src/biopsykit/questionnaires/questionnaires.py
@@ -1941,7 +1941,7 @@ def pasa(
             pasa_data[score_name + "_SelfConcept"] + pasa_data[score_name + "_ControlExp"]
         ) / 2
 
-    if all("{}_{}".format(score_name, s) in pasa_data.keys() for s in ["Primary", "Secondary"]):
+    if all("{}_{}".format(score_name, s) in pasa_data for s in ["Primary", "Secondary"]):
         pasa_data[score_name + "_StressComposite"] = (
             pasa_data[score_name + "_Primary"] - pasa_data[score_name + "_Secondary"]
         )
