### 2021-09-28 b46ea26
updated linter configuration
