+++ .prospector.yml
@@ -50,5 +50,7 @@ pylint:
     - protected-access
     - too-many-lines
     - useless-suppression
+    - unsupported-assignment-operation # because of pylint/astroid bug
+    - unsubscriptable-object # because of pylint/astroid bug
     - consider-using-f-string
 
