+++ docs/source/development/index.rst
@@ -1,11 +1,0 @@
-.. _development:
-
-===========
-Development
-===========
-
-
-.. toctree::
-    :maxdepth: 1
-
-    development_guide.md
