+++ docs/index.rst
@@ -13,7 +13,7 @@ This documentation will help you to get up to speed quickly. Take a look at the 
    Getting Started <README.md>
    guides/index.rst
    examples/index.rst
-   source/development/index.rst
+   Contributing <source/contributing/CONTRIBUTING.md>
    Changelog <CHANGELOG.md>
    Module Reference <api/modules.rst>
    About <source/about/index.rst>
