+++ docs/source/contributing/CONTRIBUTING.md
@@ -1,0 +1,364 @@
+# Contributing to BioPsyKit
+
+This document contains information for developers that want to contribute to ``BioPsyKit``.
+
+It includes further in-depth information on how to set up and use required tools
+and learn about programing methods used in the development of this project.
+
+[comment]: <> (If you are looking for a higher level overview over the guiding ideas and structure of this project, please visit the)
+[comment]: <> ([Project Structure document]&#40;project_structure.md&#41;.)
+
+## Contributing Guidelines
+
+### General Information
+
+`BioPsyKit` is under active development and welcomes everyone to contribute to code improvements, documentation, 
+testing, and new features! [Why should I contribute?](https://github.com/jonschlinkert/idiomatic-contributing/)
+
+The project is structured into different submodules that each are designed for specific purposes, such as 
+``signals``, ``sleep``, ``saliva``, ``questionnaires``, as well as overarching submodules used throughout 
+the package, such as ``plotting``, ``colors``, and ``utils``.  
+
+Before adding new features, please think about on where your features might suit best into the project structure 
+to that other users can find them easily. 
+
+Additionally, please don't forget to comment and document your code, as well as adding examples and unit 
+tests (more on that below) to your code.
+
+### Git Workflow
+
+As multiple people are expected to contribute to `BioPsyKit` at the same time, `BioPsyKit` implements a proper [Git 
+workflow](https://guides.github.com/introduction/flow/) to prevent possible issues.
+
+#### Branching structure
+
+In `BioPsyKit` we use a *main + feature branches* structure.
+This workflow is well explained [here](https://www.atlassian.com/blog/git/simple-git-workflow-is-simple).
+  
+All changes to the `main` branch should be performed using feature branches.
+Before merging, the feature branches should be rebased onto the current `main` branch.
+
+
+In summary, follow these steps:
+
+##### Create a Pull Request
+Create a new branch from the `main` branch. If you propose a new feature then name your branch something like 
+`<feature_feature-name>` where `feature-name` roughly describes the purpose of the new feature. 
+If you propose a bugfix, improvement, etc. then name your branch something like `<fix_fix-name>` where `fix-name` 
+roughly describes the bugfix or improvement you propose.
+
+Afterwards, create a [Pull Request](https://www.earthdatascience.org/courses/intro-to-earth-data-science/git-github/github-collaboration/how-to-submit-pull-requests-on-github/). 
+Be sure to have a `WIP: ` at the very beginning of its name. 
+The _source branch_ is the new branch you just created and the _target branch_ is `main`.
+
+In the description, write shortly what you are going to fix, add, or improve, then start working!
+
+##### During Development
+During development, remember to adhere to the general recommendations for feature branches. 
+
+As a reminder, feature branches...
+
+- ...should be short-lived
+- ...should be dedicated to a *single* feature
+- ...should be worked on by a *single* person
+- ...must be merged via a Pull Request and not manually
+- ...must be reviewed before merging
+- ...must pass the pipeline checks before merging
+- ...should be rebased onto `main` if possible (remember to only rebase if you are the only person working on this branch!)
+- ...should be pushed soon and often to allow everyone to see what you are working on
+- ...should be associated with a Pull Request, which is used for discussions and code review.
+- ...that are not ready to review, should have a Pull Request prefixed with `WIP: `
+- ...should also close issues that they solve, once they are merged
+
+##### Pushing Code
+
+As soon as your new contribution to `BioPsyKit` is ready, it's time to ensure your code passes all necessary checks, 
+such as formatting, linting, building documentation, and performing tests. To run these checks more easily this project 
[…]
