### 2021-09-28 mad-lab-fau/BioPsyKit#11 @richrobe
Thanks for your comments! We added guidelines for contributing to the repository starting with 0db7ab5763ed72ed608e69c9c234c45a3cfa94e9 (and the commits followed after that). Additionally, we added bug report, feature requests, and pull request templates in 4368da3474f06098fbe4d76b286c4aefab6c8810 and a88fbfd6e8c547dc551e3d3080d4746cd2080d29.
