### 2021-09-28 mad-lab-fau/BioPsyKit#11 @richrobe
Regarding your other idea: I think that's a good idea! The `stats` and `classification` modules are already briefly explained in the example Jupyter Notebooks `StatsPipeline_Plotting_Example` and `SklearnPipelinePermuter_Example`. The code for generating the Figuer 5 boxplot in the paper can be found in the `Questionnaire_Example` notebook (at the very end). We can describe the features in more detail in the respective example notebooks. Additionally, I can also add some brief examples code directly to the README?
