### 2021-09-28 fc77531
added function to print the current version of biopsykit and its core dependencies
