+++ src/biopsykit/__init__.py
@@ -1,4 +1,13 @@
 """A Python package for the analysis of biopsychological data."""
+import platform
+
+import matplotlib
+
+import numpy as np
+import pandas as pd
+import scipy
+import pingouin
+import neurokit2
 
 import biopsykit.signals  # noqa: F401
 import biopsykit.saliva  # noqa: F401
@@ -33,3 +42,26 @@ __all__ = [
 ]
 
 __version__ = "0.3.0"
+
+
+def version() -> None:
+    """Get the version of BioPsyKit and its core dependencies.
+
+    Examples
+    --------
+    >>> import biopsykit as bp
+    >>>
+    >>> bp.version()
+
+    """
+    print(
+        f"Operating System: {platform.system()} ({platform.architecture()[1]} {platform.architecture()[0]})\n",
+        f"- Python: {platform.python_version()}\n",
+        f"- BioPsyKit: {__version__}\n\n",
+        f"- NumPy: {np.__version__}\n",
+        f"- Pandas: {pd.__version__}\n",
+        f"- SciPy: {scipy.__version__}\n",
+        f"- matplotlib: {matplotlib.__version__}\n",
+        f"- NeuroKit2: {neurokit2.__version__}\n",
+        f"- pingouin: {pingouin.__version__}\n",
+    )
