### 2021-09-28 74e72d7
updated example data, added new example data: get_hr_subject_data_dict_example()
