### 2021-09-28 01b6069
added tests for biopsykit.example_data
