### 2021-09-28 mad-lab-fau/BioPsyKit#10 @richrobe
Thank you for pointing that out! We assume this is due to the fact that the example data is not included in the package when installing it from PyPi.

We will update our code so that it will first check whether the package was installed _manually_, by cloning the git repository, then all example data should be downloaded and the examples should work correctly, or via _PyPi_. In the latter case the example data is not contained and we will download the example data similar to how scikit-learn is handling this (see https://github.com/scikit-learn/scikit-learn/blob/844b4be24/sklearn/datasets/_base.py).

We also updated the minimal working examples in the README, which still contained some bugs, in 6935600009cbc143ef4fd750211c249b9ba09d1e.
