### 2021-09-29 252f33c
Example data is now downloaded from GitHub into a local user folder if the package was installed from PyPi instead of manually (repository checked out from GitHub)
