+++ src/biopsykit/example_data.py
@@ -1,7 +1,12 @@
-"""Module providing functions to load different example data."""
-import re
-from pathlib import Path
+"""Module providing functions to load different example data.
+
+The data is either taken from the local file system in case biopsykit was installed manually or the example data is
+downloaded into the local user folder.
+"""
 from typing import Sequence, Dict, Optional, Tuple, Union
+from urllib.request import urlretrieve
+from pathlib import Path
+from tqdm.auto import tqdm
 
 import pandas as pd
 
@@ -15,6 +20,7 @@ from biopsykit.utils.datatype_helper import (
     SleepEndpointDataFrame,
     _SalivaMeanSeDataFrame,
 )
+from biopsykit.utils.file_handling import mkdirs
 from biopsykit.io import load_subject_condition_list, load_time_log, load_questionnaire_data
 from biopsykit.io.carwatch_logs import load_log_one_subject
 from biopsykit.io.eeg import load_eeg_raw_muse
@@ -25,7 +31,9 @@ from biopsykit.io.sleep_analyzer import load_withings_sleep_analyzer_raw_folder
 from biopsykit.io.sleep_analyzer import load_withings_sleep_analyzer_summary
 from biopsykit.utils._types import path_t
 
-_EXAMPLE_DATA_PATH = Path(__file__).parent.parent.parent.joinpath("example_data")
+_EXAMPLE_DATA_PATH_LOCAL = Path(__file__).parent.parent.parent.joinpath("example_data")
+_EXAMPLE_DATA_PATH_HOME = Path.home().joinpath(".biopsykit_data")
+_REMOTE_DATA_PATH = "https://raw.githubusercontent.com/mad-lab-fau/BioPsyKit/main/example_data/"
 
 __all__ = [
     "get_file_path",
@@ -45,6 +53,64 @@ __all__ = [
 ]
 
 
+def _is_installed_manually() -> bool:
+    """Check whether biopsykit was installed manually and example data exists in the local path.
+
+    Returns
+    -------
+    bool
+        ``True`` if biopsykit was installed manually, ``False`` otherwise
+
+    """
+    return (_EXAMPLE_DATA_PATH_LOCAL / "__init__.py").is_file()
+
+
+def _get_data(file_name: str) -> path_t:
+    if _is_installed_manually():
+        return _EXAMPLE_DATA_PATH_LOCAL.joinpath(file_name)
+    path = _EXAMPLE_DATA_PATH_HOME.joinpath(file_name)
+    if path.exists():
+        return path
+    mkdirs(path.parent)
+    return _fetch_from_remote(file_name, path)
+
+
+def _fetch_from_remote(file_name: str, file_path: path_t) -> path_t:
+    """Download remote dataset (helper function).
+
+    Parameters
+    ----------
+    file_name : str
+        file name
+    file_path : str
+        path to file
+
+    Returns
+    -------
+    :class:`~pathlib.Path`
+        path to downloaded file
+
+    """
+    url = _REMOTE_DATA_PATH + file_name
+    print(f"Downloading file {file_name} from remote URL: {url}.")
+    with tqdm(unit="B", unit_scale=True, unit_divisor=1024, miniters=1, desc=file_name) as t:
+        urlretrieve(url, filename=file_path, reporthook=_tqdm_hook(t))
+    return file_path
+
+
+def _tqdm_hook(t):
+    """Wrap tqdm instance."""
+    last_b = [0]
+
+    def update_to(b=1, bsize=1, tsize=None):
+        if tsize is not None:
[…]
