### 2021-09-30 511f878
updated example data to include all required data for the example notebooks
