+++ tests/test_example_data/test_example_data.py
@@ -8,7 +8,7 @@ from biopsykit.example_data import *
 from biopsykit.example_data import get_eeg_example, get_questionnaire_example_wrong_range
 from biopsykit.questionnaires import pss
 from biopsykit.questionnaires.utils import find_cols
-from biopsykit.utils._datatype_validation_helper import _assert_is_dtype, _assert_has_columns
+from biopsykit.utils._datatype_validation_helper import _assert_is_dtype, _assert_has_columns, _assert_has_index_levels
 from biopsykit.utils.datatype_helper import (
     is_subject_condition_dataframe,
     is_saliva_raw_dataframe,
@@ -25,6 +25,9 @@ from biopsykit.utils.exceptions import ValueRangeError
 @contextmanager
 def does_not_raise():
     yield
+
+
+# TODO add new Test Class with pre-and post-test hook that allows to test example data downloading (by modifying example_data.__init__() so that it can't be found anymore, pretending the package wasn't installed manually and data is downloaded from remote)
 
 
 class TestExampleData:
@@ -70,12 +73,28 @@ class TestExampleData:
         data = get_saliva_example()
         is_saliva_raw_dataframe(data, "cortisol")
 
+    def test_get_saliva_example_plate_format(self):
+        data = get_saliva_example_plate_format()
+        _assert_is_dtype(data, pd.DataFrame)
+
     def test_get_saliva_mean_se_example(self):
         data = get_saliva_mean_se_example()
         _assert_is_dtype(data, dict)
         for key, df in data.items():
             assert key in ["cortisol", "amylase", "il6"]
             is_saliva_mean_se_dataframe(df)
+
+    def test_get_hr_result_sample(self):
+        data = get_hr_result_sample()
+        _assert_is_dtype(data, pd.DataFrame)
+        _assert_has_columns(data, [["HR"]])
+
+    def test_get_hr_ensemble_sample(self):
+        data = get_hr_ensemble_sample()
+        _assert_is_dtype(data, dict)
+        for key, df in data.items():
+            _assert_is_dtype(key, str)
+            _assert_is_dtype(df, pd.DataFrame)
 
     def test_get_mist_hr_example(self):
         data = get_mist_hr_example()
@@ -84,6 +103,22 @@ class TestExampleData:
     def test_get_hr_subject_data_dict_example(self):
         data = get_hr_subject_data_dict_example()
         is_hr_subject_data_dict(data)
+
+    def test_get_ecg_path_example(self):
+        path = get_ecg_path_example()
+        assert path.exists()
+
+    def test_get_car_watch_log_path_example(self):
+        path = get_car_watch_log_path_example()
+        assert path.exists()
+
+    def test_get_car_watch_log_data_zip_path_example(self):
+        path = get_car_watch_log_data_zip_path_example()
+        assert path.exists()
+
+    def test_get_car_watch_log_path_all_subjects_example(self):
+        path = get_car_watch_log_path_all_subjects_example()
+        assert path.exists()
 
     def test_get_ecg_example(self):
         data, fs = get_ecg_example()
@@ -97,10 +132,60 @@ class TestExampleData:
         assert fs == 256.0
 
     @pytest.mark.parametrize(
+        "data_source, expected",
+        [
+            ("heart_rate", does_not_raise()),
+            ("respiration_rate", does_not_raise()),
+            ("sleep_state", does_not_raise()),
+            ("snoring", does_not_raise()),
+            ("data", pytest.raises(ValueError)),
+        ],
+    )
+    def test_get_sleep_analyzer_raw_file_unformatted_raises(self, data_source, expected):
+        with expected:
+            get_sleep_analyzer_raw_file_unformatted(data_source)
[…]
