+++ src/biopsykit/example_data.py
@@ -21,13 +21,23 @@ from biopsykit.utils.datatype_helper import (
     _SalivaMeanSeDataFrame,
 )
 from biopsykit.utils.file_handling import mkdirs
-from biopsykit.io import load_subject_condition_list, load_time_log, load_questionnaire_data
+from biopsykit.io import (
+    load_subject_condition_list,
+    load_time_log,
+    load_questionnaire_data,
+    load_long_format_csv,
+    load_pandas_dict_excel,
+)
 from biopsykit.io.carwatch_logs import load_log_one_subject
 from biopsykit.io.eeg import load_eeg_raw_muse
 from biopsykit.io.ecg import load_hr_phase_dict
 from biopsykit.io.nilspod import load_dataset_nilspod
-from biopsykit.io.saliva import load_saliva_wide_format
-from biopsykit.io.sleep_analyzer import load_withings_sleep_analyzer_raw_folder
+from biopsykit.io.saliva import load_saliva_wide_format, load_saliva_plate
+from biopsykit.io.sleep_analyzer import (
+    load_withings_sleep_analyzer_raw_folder,
+    WITHINGS_RAW_DATA_SOURCES,
+    load_withings_sleep_analyzer_raw_file,
+)
 from biopsykit.io.sleep_analyzer import load_withings_sleep_analyzer_summary
 from biopsykit.utils._types import path_t
 
@@ -38,19 +48,30 @@ _REMOTE_DATA_PATH = "https://raw.githubusercontent.com/mad-lab-fau/BioPsyKit/mai
 __all__ = [
     "get_file_path",
     "get_condition_list_example",
+    "get_saliva_example_plate_format",
     "get_saliva_example",
     "get_saliva_mean_se_example",
     "get_mist_hr_example",
+    "get_hr_result_sample",
+    "get_hr_ensemble_sample",
     "get_hr_subject_data_dict_example",
+    "get_ecg_path_example",
     "get_ecg_example",
     "get_ecg_example_02",
+    "get_sleep_analyzer_raw_file_unformatted",
+    "get_sleep_analyzer_raw_file",
     "get_sleep_analyzer_raw_example",
     "get_sleep_analyzer_summary_example",
     "get_sleep_imu_example",
+    "get_car_watch_log_path_example",
+    "get_car_watch_log_data_zip_path_example",
+    "get_car_watch_log_path_all_subjects_example",
     "get_car_watch_log_data_example",
     "get_time_log_example",
     "get_questionnaire_example",
 ]
+
+# TODO add SHA256 check to assert whether remote example data was changed and should be re-downloaded.
 
 
 def _is_installed_manually() -> bool:
@@ -148,6 +169,56 @@ def get_condition_list_example() -> SubjectConditionDataFrame:
     )
 
 
+def get_saliva_example_plate_format(
+    sample_id_col: Optional[str] = None,
+    data_col: Optional[str] = None,
+    id_col_names: Optional[Sequence[str]] = None,
+    regex_str: Optional[str] = None,
+    sample_times: Optional[Sequence[int]] = None,
+    condition_list: Optional[Union[Sequence, Dict[str, Sequence], pd.Index]] = None,
+) -> pd.DataFrame:
+    r"""Return example saliva data from "plate" format.
+
+    Parameters
+    ----------
+    sample_id_col: str, optional
+        column name of the Excel sheet containing the sample ID. Default: "sample ID"
+    data_col: str, optional
+        column name of the Excel sheet containing saliva data to be analyzed.
+        Default: Select default column name based on ``biomarker_type``, e.g. ``cortisol`` => ``cortisol (nmol/l)``
+    id_col_names: list of str, optional
[…]
