### 2021-09-30 eef4b47
added quick examples for "stats" and "classification"
