### 2021-09-30 92cbbe8
bp.classification: addressed numpy deprecation warnings
