### 2021-09-30 a5b11d7
- improved colors submodule by providing an easier function interface - added example notebook to demonstrate use of colors submodule
