+++ src/biopsykit/colors/colors.py
@@ -2,14 +2,17 @@
 
 For most plots, BioPsyKit uses FAU's colors scheme.
 """
-from typing import Union, Sequence, Optional, Dict
+from typing import Sequence, Optional, Dict, Callable
 
 from colorsys import rgb_to_hls, hls_to_rgb
+from typing_extensions import Literal, get_args
+
 from matplotlib.colors import to_hex, to_rgb
 import seaborn as sns
 
 __all__ = [
     "FAU_COLORS",
+    "fau_palette_by_name",
     "fau_color_dict",
     "fau_palette",
     "fau_palette_blue",
@@ -22,7 +25,7 @@ __all__ = [
     "adjust_color",
 ]
 
-FAU_COLORS = ["fau", "tech", "phil", "med", "nat", "wiso"]
+FAU_COLORS = Literal["fau", "tech", "phil", "med", "nat", "wiso"]
 """
 Available color keys.
 """
@@ -43,22 +46,40 @@ fau_palette = sns.color_palette(fau_color_dict.values())  #: :meta hide-value:
 """FAU color palette that can be used with seaborn and matplotlib."""
 
 
-def fau_palette_blue(palette_type: Optional[str] = None) -> Sequence[str]:
+def fau_palette_by_name(name: FAU_COLORS) -> Callable:
+    """Return the function to create a FAU color palette by the name the color.
+
+    Parameters
+    ----------
+    name : str
+        Color name. Must be one of :const:`biopsykit.colors.FAU_COLORS`.
+
+    Returns
+    -------
+    function
+        function to create FAU color palette with
+
+    Examples
+    --------
+    >>> from biopsykit.colors import fau_palette_by_name
+    >>> fau_palette_by_name("tech")
+    <function biopsykit.colors.colors.fau_palette_tech()>
+
+    """
+    if name not in get_args(FAU_COLORS):
+        raise ValueError("'name' must be one of {}, got '{}'!".format(get_args(FAU_COLORS), name))
+    if name == "fau":
+        name = "blue"
+    return globals()["fau_palette_{}".format(name)]
+
+
+def fau_palette_blue(n_colors: Optional[int] = 8) -> Sequence[str]:
     """Return a seaborn palette with fau-blue color nuances.
 
-    By default, a palette with 10 color nuances is generated.
-    Using the ``palette_type`` parameter specialized palettes can be generated.
-
     Parameters
     ----------
-    palette_type : str, optional
-        Specify specialized color palette (or ``None`` to return default palette). Default: ``None``.
-        Available palette types:
-
-        * ``line_2``: For line plots with two elements
-        * ``line_3``: For line plots with three elements
-        * ``ensemble_3``: For ensemble plots (mean ± std) with three elements
-        * ``box_2``: For boxplots with two elements
+    n_colors : int
+        number of colors in the palette. Default: 8
 
     Returns
     -------
@@ -66,250 +87,22 @@ def fau_palette_blue(palette_type: Optional[str] = None) -> Sequence[str]:
         list of RGB tuples
 
     """
-    # generated using this link: https://noeldelgado.github.io/shadowlord
-    fau_blue = sns.color_palette(
-        [
-            "#001628",
-            "#001c33",
-            "#002747",
-            "#002d51",
-            "#003865",
-            "#194c74",
-            "#336084",
-            "#4d7493",
-            "#809cb2",
-            "#b3c3d1",
-            "#e6ebf0",
-        ]
-    )
-    if palette_type == "line_3":
-        return [fau_blue[4], fau_blue[7], fau_blue[9]]
-    if palette_type == "ensemble":
-        return fau_blue[1::2]
-    if palette_type == "box_2":
-        return fau_blue[5::4]
-    if palette_type == "line_2":
-        return fau_blue[2::5]
[…]
