+++ src/biopsykit/colors/__init__.py
@@ -6,6 +6,7 @@ from biopsykit.colors.colors import (
     FAU_COLORS,
     fau_color_dict,
     fau_palette,
+    fau_palette_by_name,
     fau_palette_blue,
     fau_palette_phil,
     fau_palette_med,
@@ -20,6 +21,7 @@ __all__ = [
     "FAU_COLORS",
     "fau_color_dict",
     "fau_palette",
+    "fau_palette_by_name",
     "fau_palette_blue",
     "fau_palette_phil",
     "fau_palette_med",
