### 2021-09-30 015f005
formatting and doc improvements
