+++ src/biopsykit/utils/data_processing.py
@@ -627,152 +627,6 @@ def _splits_subject_conditions(data_dict: Dict[str, Any], subject_list: Sequence
     return {key: _splits_subject_conditions(value, subject_list) for key, value in data_dict.items()}
 
 
-# def param_subphases(
-#     ecg_processor: Optional["EcgProcessor"] = None,
-#     dict_ecg: Optional[Dict[str, pd.DataFrame]] = None,
-#     dict_rpeaks: Optional[Dict[str, pd.DataFrame]] = None,
-#     subphases: Optional[Sequence[str]] = None,
-#     subphase_durations: Optional[Sequence[int]] = None,
-#     param_types: Optional[Union[str, Sequence[str]]] = "all",
-#     sampling_rate: Optional[int] = 256,
-#     include_total: Optional[bool] = True,
-#     title: Optional[str] = None,
-# ) -> pd.DataFrame:
-#     """
-#     Computes specified parameters (HRV / RSA / ...) over phases and subphases **for one subject**.
-#
-#     To use this function, either simply pass an ``EcgProcessor`` object or two dictionaries
-#     ``dict_ecg`` and ``dict_rpeaks`` resulting from ``EcgProcessor.ecg_process()``.
-#
-#     Parameters
-#     ----------
-#     ecg_processor : EcgProcessor, optional
-#         `EcgProcessor` object
-#     dict_ecg : dict, optional
-#         dict with dataframes of processed ECG signals. Output from `EcgProcessor.ecg_process()`.
-#     dict_rpeaks : dict, optional
-#         dict with dataframes of processed R peaks. Output from `EcgProcessor.ecg_process()`.
-#     subphases : list of int
-#         list of subphase names
-#     subphase_durations : list of str
-#         list of subphase durations
-#     param_types : list or str, optional
-#         list with parameter types to compute or 'all' to compute all available parameters. Choose from a subset of
-#         ['hrv', 'rsa'] to compute HRV and RSA parameters, respectively.
-#     sampling_rate : float, optional
-#         Sampling rate of recorded data. Not needed if ``ecg_processor`` is supplied as parameter. Default: 256 Hz
-#     include_total : bool, optional
-#         ``True`` to also compute parameters over the complete phases (in addition to only over subphases),
-#         ``False`` to only compute parameters over the single subphases. Default: ``True``
-#     title : str, optional
-#         Optional title of the processing progress bar. Default: ``None``
-#
-#     Returns
-#     -------
-#     pd.DataFrame
-#         dataframe with computed parameters over the single subphases
-#     """
-#     import biopsykit.signals.ecg as ecg
-#
-#     if ecg_processor is None and dict_rpeaks is None and dict_ecg is None:
-#         raise ValueError("Either `ecg_processor` or `dict_rpeaks` and `dict_ecg` must be passed as arguments!")
-#
-#     if subphases is None or subphase_durations is None:
-#         raise ValueError("Both `subphases` and `subphase_durations` are required as parameter!")
-#
-#     # TODO change
-#     # get all desired parameter types
-#     possible_param_types = {
-#         "hrv": ecg.EcgProcessor.hrv_process,
-#         # "rsp": ecg.EcgProcessor.rsp_rsa_process,
-#     }
-#     if param_types == "all":
-#         param_types = possible_param_types
-#
-#     if isinstance(param_types, str):
-#         param_types = {param_types: possible_param_types[param_types]}
-#     if not all([param in possible_param_types for param in param_types]):
-#         raise ValueError(
-#             "`param_types` must all be of {}, not {}".format(possible_param_types.keys(), param_types.keys())
-#         )
-#
[…]
