### 2021-09-30 a95755e
added stats example data
