### 2021-09-30 9c508d6
changed interface to bp.io.load_time_log() to match the interface of other data loader functions (e.g., bp.io.load_questionnaire_data())
