+++ src/biopsykit/io/__init__.py
@@ -1,17 +1,19 @@
 """Module providing input/output functions."""
 from biopsykit.io.io import (
-    write_pandas_dict_excel,
-    write_result_dict,
+    load_long_format_csv,
     load_questionnaire_data,
     load_time_log,
     load_subject_condition_list,
     load_pandas_dict_excel,
     load_codebook,
+    write_pandas_dict_excel,
+    write_result_dict,
     convert_time_log_datetime,
 )
 from biopsykit.io import ecg, eeg, nilspod, saliva, sleep, sleep_analyzer
 
 __all__ = [
+    "load_long_format_csv",
     "load_time_log",
     "load_subject_condition_list",
     "load_questionnaire_data",
