+++ tests/test_io/test_io_io.py
@@ -144,74 +144,109 @@ def result_dict_correct():
 
 class TestIoIo:
     @pytest.mark.parametrize(
-        "file_path, continuous_time, index_cols, phase_cols, expected",
+        "file_path, continuous_time, subject_col, condition_col, additional_index_cols, phase_cols, expected",
         [
-            ("time_log.csv", True, None, None, does_not_raise()),
-            ("time_log.xlsx", True, None, None, does_not_raise()),
-            ("time_log.csv", True, ["time"], None, pytest.raises(ValidationError)),
-            ("time_log.csv", True, "subject", None, does_not_raise()),
-            ("time_log.csv", True, ["subject", "time"], None, pytest.raises(ValidationError)),
-            ("time_log.csv", True, None, ["Baseline", "Intervention", "Stress", "Recovery"], does_not_raise()),
+            ("time_log.csv", True, None, None, None, None, does_not_raise()),
+            ("time_log.xlsx", True, None, None, None, None, does_not_raise()),
+            ("time_log.csv", True, "time", None, None, None, pytest.raises(ValidationError)),
+            ("time_log.csv", True, "subject", None, None, None, does_not_raise()),
+            ("time_log.csv", True, "subject", "condition", None, None, does_not_raise()),
+            ("time_log.csv", True, ["subject"], None, None, None, pytest.raises(ValidationError)),
+            ("time_log.csv", True, "subject", ["condition"], None, None, pytest.raises(ValidationError)),
+            ("time_log.csv", True, "subject", None, ["time"], None, pytest.raises(ValidationError)),
             (
                 "time_log.csv",
                 True,
                 None,
+                None,
+                None,
+                ["Baseline", "Intervention", "Stress", "Recovery"],
+                does_not_raise(),
+            ),
+            (
+                "time_log.csv",
+                True,
+                None,
+                None,
+                None,
                 ["Baseline", "Intervention", "Stress", "Postline"],
                 pytest.raises(ValidationError),
             ),
-            ("time_log_other_index_names.csv", True, None, None, does_not_raise()),
-            ("time_log_other_index_names.csv", True, ["subject", "condition"], None, pytest.raises(ValidationError)),
+            ("time_log_other_index_names.csv", True, None, None, None, None, pytest.raises(ValidationError)),
             (
                 "time_log_other_index_names.csv",
                 True,
-                {"ID": "subject", "Condition": "condition"},
+                "subject",
+                "condition",
+                None,
+                None,
+                pytest.raises(ValidationError),
+            ),
+            (
+                "time_log_other_index_names.csv",
+                True,
+                "ID",
+                "Condition",
+                None,
                 None,
                 does_not_raise(),
             ),
-            ("time_log_not_continuous.csv", True, None, None, does_not_raise()),
-            ("time_log_not_continuous.csv", False, None, None, does_not_raise()),
-            ("time_log.xlsx", False, None, None, pytest.raises(ValidationError)),
-            ("time_log_wrong_column_names.csv", False, None, None, pytest.raises(ValidationError)),
+            ("time_log_not_continuous.csv", True, None, None, None, None, does_not_raise()),
+            ("time_log_not_continuous.csv", False, None, None, None, None, does_not_raise()),
+            ("time_log.xlsx", False, None, None, None, None, pytest.raises(ValidationError)),
+            ("time_log_wrong_column_names.csv", False, None, None, None, None, pytest.raises(ValidationError)),
         ],
     )
-    def test_load_time_log_raises(self, file_path, continuous_time, index_cols, phase_cols, expected):
+    def test_load_time_log_raises(
+        self, file_path, continuous_time, subject_col, condition_col, additional_index_cols, phase_cols, expected
+    ):
         with expected:
             load_time_log(
                 file_path=TEST_FILE_PATH.joinpath(file_path),
[…]
