+++ src/biopsykit/io/io.py
@@ -1,7 +1,7 @@
 """Module containing different I/O functions to load time log data, subject condition lists, questionnaire data, etc."""
 
 from pathlib import Path
-from typing import Optional, Union, Sequence, Dict
+from typing import Optional, Union, Sequence, Dict, Tuple
 
 import datetime
 import pytz
@@ -43,7 +43,9 @@ __all__ = [
 
 def load_time_log(
     file_path: path_t,
-    index_cols: Optional[Union[str, Sequence[str], Dict[str, str]]] = None,
+    subject_col: Optional[str] = None,
+    condition_col: Optional[str] = None,
+    additional_index_cols: Optional[Union[str, Sequence[str]]] = None,
     phase_cols: Optional[Union[Sequence[str], Dict[str, str]]] = None,
     continuous_time: Optional[bool] = True,
     **kwargs,
@@ -57,11 +59,20 @@ def load_time_log(
     ----------
     file_path : :class:`~pathlib.Path` or str
         path to time log file. Must either be an Excel or csv file
-    index_cols : str, list of str or dict, optional
-        specifies dataframe index. Can either be a string or a list of strings to indicate column name(s) from the
-        dataframe that should be used as index level(s), or ``None`` for no index.
-        If the index levels of the time log dataframe should have different names than the columns in the file,
-        a dict specifying the mapping (column_name : new_index_name) can be passed. Default: ``None``
+    subject_col : str, optional
+        name of column containing subject IDs or ``None`` to use default column name ``subject``.
+        According to BioPsyKit's convention, the subject ID column is expected to have the name ``subject``.
+        If the subject ID column in the file has another name, the column will be renamed in the dataframe
+        returned by this function.
+    condition_col : str, optional
+        name of column containing condition assignments or ``None`` to use default column name ``condition``.
+        According to BioPsyKit's convention, the condition column is expected to have the name ``condition``.
+        If the condition column in the file has another name, the column will be renamed in the dataframe
+        returned by this function.
+    additional_index_cols : str, list of str, optional
+        additional index levels to be added to the dataframe.
+        Can either be a string or a list strings to indicate column name(s) that should be used as index level(s),
+        or ``None`` for no additional index levels. Default: ``None``
     phase_cols : list of str or dict, optional
         list of column names that contain time log information or ``None`` to use all columns.
         If the column names of the time log dataframe should have different names than the columns in the file,
@@ -95,21 +106,22 @@ def load_time_log(
     >>> # load time log file into a pandas dataframe
     >>> data = bp.io.load_time_log(file_path)
     >>> # Example 2:
-    >>> # load time log file into a pandas dataframe and specify the 'subject_id' column
-    >>> # in the time log file to be the index of the dataframe
-    >>> data = bp.io.load_time_log(file_path, index_cols="subject_id")
+    >>> # load time log file into a pandas dataframe and specify the "ID" column
+    >>> # (instead of the default "subject" column) in the time log file to be the index of the dataframe
+    >>> data = bp.io.load_time_log(file_path, subject_col="ID")
     >>> # Example 3:
-    >>> # load time log file into a pandas dataframe, specify the columns "subject_id" and "condition"
-    >>> # to be the new index, and the columns 'Phase1' 'Phase2' and 'Phase3' to be used
-    >>> # for extracting time information
[…]
