### 2021-09-30 0bd9b59
bp.plotting: renamed "colormap" parameter in all plotting functions to "palette"
