### 2021-09-30 768f1a6
bp.protocols: added "export_hr_ensemble()" function and smaller fixes
