### 2021-09-30 3ac516f
bp.utils.dataframe_handling.multi_xs() now also supports Series objects as input
