### 2021-09-30 85f24ea
bp.stats.StatsPipeline now supports rounding p-values
