### 2021-09-30 67ac96d
bp.stats.StatsPipeline: improved interface of multicomp()
