### 2021-09-30 791c83b
bp.utils.data_processing: small improvements: - mean_se_per_phase(): fixed order of group keys when computing mean and standard error (keys were previously sorted by name when grouping) - exclude_subjects(): improved warning message displayed when subject(s) are not present in dataset - split_dict_into_subphases(): now also supports dataframes with DateTimeIndex to be split according to their time index
