### 2021-09-30 mad-lab-fau/BioPsyKit#11 @richrobe
We updated the README to include the minimal example codes in b00bc44e6834021a1c97e987daf58f67c341d776!
