### 2021-09-30 mad-lab-fau/BioPsyKit#14 @richrobe
Thank you for your comments! We updated the example notebooks in 1725a2706f730aedd3328febd0f7c9cbd6b0e174.
