### 2021-10-04 mad-lab-fau/BioPsyKit#10 @zen-juen
Hi @richrobe that's great! I think this new version is not yet updated on PyPI though, so I can't review the changes as of now
