+++ src/biopsykit/__init__.py
@@ -41,7 +41,7 @@ __all__ = [
     "utils",
 ]
 
-__version__ = "0.3.0"
+__version__ = "0.3.1"
 
 
 def version() -> None:
