+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "biopsykit"
-version = "0.3.0"
+version = "0.3.1"
 description = "A Python package for the analysis of biopsychological data."
 authors = [
     "Robert Richer <robert.richer@fau.de>",
