+++ paper/paper.md
@@ -35,33 +35,36 @@ bibliography: paper.bib
 ---
 
 # Summary
-
-Biopsychology is a field of psychology that analyzes how biological processes interact with behaviour, emotion, 
+_Biopsychology_ is a field of psychology that analyzes how biological processes interact with behaviour, emotion, 
 cognition, and other mental processes. Biopsychology covers, among others, the topics of sensation and perception, 
 emotion regulation, movement (and control of such), sleep and biological rhythms, as well as acute and chronic stress.
 
-To assess the interaction between biological and mental processes a variety of different methods are used in the 
-field of biopsychology, such as _electrophysiology_, assessed via biosignals like electrocardiogram (ECG), 
-electrodermal activity (EDA), electromyogram (EMG), or electroencephalogram (EEG), _neuroendocrine and inflammatory 
-biomarker_, assessed via saliva- and blood-based samples, _self-reports_, assessed via psychological questionnaires, 
-as well as _sleep, activity and movement_, assessed via inertial measurement units (IMUs).
+To assess the interaction between biological and mental processes a variety of different modalities are used in the 
+field of biopsychology, such as _electrophysiology_, assessed, for instance, via electrocardiography (ECG), 
+electrodermal activity (EDA), or electroencephalography (EEG), _sleep, activity and movement_, assessed via inertial measurement units (IMUs), _neuroendocrine and inflammatory 
+biomarker_, assessed by saliva and blood samples, as well as _self-reports_, assessed via psychological questionnaires. 
 
-These methods are used to collect data either during standardized procedures in the laboratory or in the wild. 
+These different modalities are collected either "in the lab", during standardized laboratory protocols, or "in the wild", during unsupervised protocols in home environments. 
 The collected data are typically analyzed using statistical methods, or, more recently, using machine learning methods.
 
-In order to combine all these tools necessary for a researcher in the field of biopsychology into 
-one single Python package we developed `BioPsyKit`.
+
+While some software packages exist that allow for the analysis of single data modalities, such as electrophysiological data, or sleep, activity and movement data, no packages are available for the analysis of other modalities, such as neuroendocrine and inflammatory biomarker, and self-reports. In order to fill this gap, and, simultaneously, to combine all required tools analyzing biopsychological data from beginning to end into one single Python package, we developed `BioPsyKit`.
 
 
 # Statement of need
-Researchers in biopsychology often combine different assessment modalities during experiments in order to capture the 
-interaction between biological and mental processes. One example might be collecting salivary biomarker 
-(e.g., cortisol) during an acute stress protocol and investigating the correlation between biomarker and psychometric 
-data assessed via self-reports, such as perceived stress, state anxiety, or positive/negative affect. 
-However, currently, there exist no Python package that allows to systematically combine, process, and analyze these 
-different data modalities out of one hand by using one common API. For that reason `BioPsyKit` enables researchers to 
-write cleaner and reproducible analysis code, export analysis results in a standardized format, and create high 
-quality figures with for scientific publications. 
+Researchers in biopsychology often face the challenge to analyze data from different assessment modalities during experiments in order to capture the 
+complex interaction between biological and mental processes. 
+
+One example might be collecting (electro)physiological (e.g., ECG) or 
+salivary biomarkers (e.g., cortisol) during an acute stress protocol, and investigating the correlation 
+between biomarkers and psychometric data assessed via self-reports, such as perceived stress, state anxiety, 
+or positive/negative affect. Another example is the assessment of relationships between sleep and neuroendocrine responses in the morning. To assess the 
[…]
