+++ paper/paper.bib
@@ -1,58 +1,16 @@
-@article{Makowski2021,
-  title = {{{NeuroKit2}}: {{A Python}} Toolbox for Neurophysiological Signal Processing},
-  shorttitle = {{{NeuroKit2}}},
-  author = {Makowski, Dominique and Pham, Tam and Lau, Zen J. and Brammer, Jan C. and Lespinasse, Fran{\c c}ois and Pham, Hung and Sch{\"o}lzel, Christopher and Chen, S. H. Annabel},
-  year = {2021},
-  month = aug,
-  journal = {Behavior Research Methods},
-  volume = {53},
-  number = {4},
-  pages = {1689--1696},
-  issn = {1554-3528},
-  doi = {10.3758/s13428-020-01516-y},
-  language = {en},
-}
 
-@article{Cole1992,
-  title = {Automatic {{Sleep}}/{{Wake Identification From Wrist Activity}}},
-  author = {Cole, R. J. and Kripke, D. F. and Gruen, W. and Mullaney, D. J. and Gillin, J. C.},
-  year = {1992},
-  journal = {Sleep},
-  volume = {15},
-  number = {5},
-  pages = {461--469},
-  issn = {01618105},
-  doi = {10.1093/sleep/15.5.461},
-  pmid = {1455130},
-  keywords = {Sleep/Wake},
-}
-
-@article{Sadeh1994,
-  title = {Activity-{{Based Sleep}}-{{Wake Identification}}: {{An Empirical Test}} of {{Methodological Issues}}},
-  author = {Sadeh, A. and Sharkey, K. M. and Carskadon, M. A.},
-  year = {1994},
-  journal = {Sleep},
-  volume = {17},
-  number = {3},
-  pages = {201--207},
-  issn = {01618105},
-  doi = {10.1093/sleep/17.3.201},
-  pmid = {7939118},
-  keywords = {Actigraph,Sleep/Wake},
-}
-
-@article{Pruessner2003,
-  title = {Two Formulas for Computation of the Area under the Curve Represent Measures of Total Hormone Concentration versus Time-Dependent Change},
-  author = {Pruessner, Jens C. and Kirschbaum, Clemens and Meinlschmid, Gunther and Hellhammer, Dirk H.},
-  year = {2003},
-  journal = {Psychoneuroendocrinology},
-  volume = {28},
-  number = {7},
-  pages = {916--931},
-  issn = {03064530},
-  doi = {10.1016/S0306-4530(02)00108-7},
-  abstract = {Study protocols in endocrinological research and the neurosciences often employ repeated measurements over time to record changes in physiological or endocrinological variables. While it is desirable to acquire repeated measurements for finding individual and group differences with regard to response time and duration, the amount of data gathered often represents a problem for the statistical analysis. When trying to detect possible associations between repeated measures and other variables, the area under the curve (AUC) is routinely used to incorporate multiple time points. However, formulas for computation of the AUC are not standardized across laboratories, and existing differences are usually not presented when discussing results, thus causing possible variability, or incompatibility of findings between research groups. In this paper, two formulas for calculation of the area under the curve are presented, which are derived from the trapezoid formula. These formulas are termed 'Area under the curve with respect to increase' (AUCI) and 'Area under the curve with respect to ground' (AUCG). The different information that can be derived from repeated measurements with these two formulas is exemplified using artificial and real data from recent studies of the authors. It is shown that depending on which formula is used, different associations with other variables may emerge. Consequently, it is recommended to employ both formulas when analyzing data sets with repeated measures.},
-  pmid = {12892658},
+@article{Christakis2019,
[…]
