### 2021-10-06 #3702 @richrobe
Dear all (@osorensen, @espenhgn, @zen-juen),

I have incorporated all proposed changes into the BioPsyKit package as well as into the paper. The changes are also published in the [0.3.1](https://github.com/mad-lab-fau/BioPsyKit/releases/tag/0.3.1) release, the paper can be found in the [joss](https://github.com/mad-lab-fau/BioPsyKit/tree/joss/paper) branch.

Let me know what you think!

Best,
Robert
