### 2021-10-07 #3702 @zen-juen
Hi @richrobe!

Thanks for addressing the comments raised! My thoughts sectioned as follows:

- **Code/Documentation**
The guidelines you've added for contributing/PR/raising issues in the repo look really thorough and comprehensive, great job! I can load and run the example datasets nicely now, with the exception of `bp.example_data.get_condition_list_example()` in the saliva example notebook. It returns a similar error of not being able to locate the file, so it appears that you’ll need to allow for downloading of `condition_list.csv` too. I also see that the documentation page is working now, but the example notebooks are not displayed yet. I don't think this is specifically necessary for acceptance at this juncture but some other things you might want to consider is to enable displaying of plots and results in the jupyter notebooks when they are up and running. Additionally,  it'll also be nice to have more information in [stats pipeline example](https://github.com/mad-lab-fau/BioPsyKit/blob/main/examples/SklearnPipelinePermuter_Example.ipynb) (i.e., about the datasets, and the several parameters used in the code). Other than these points, I'm happy with the documentation and the overall amount of guidance provided to navigate the package.

Another small design-related issue is: I'm wondering if when preparing the dataset for analyses across multiple phases (such as in https://github.com/mad-lab-fau/BioPsyKit/blob/main/examples/ECG_Processing_Example.ipynb), do users have to specifically pass exact time points that define the phase? In case that information is not available, is it possible to distinguish phases by a list of durations instead?

- **Paper**
You've addressed very nicely the additional points about the gap that `BioPsyKit` fills, its distinction from other related packages, and the purpose of the protocols. The figure depicting the package's structure is also much clearer now. I have no other comments other than I think it'll be helpful for interested users to know how the figures were obtained (by referring them to the respective example notebook links that generated them).
