### 2021-10-08 12ce67d
updated external intersphinx directories
