### 2021-10-08 a0771f4
fixed formatting issue in function doc
