### 2021-10-08 03b02f5
fixed loading example condition list
