+++ dodo.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import platform
 
 DOIT_CONFIG = {
-    "default_tasks": ["format", "test", "lint"],
+    "default_tasks": ["format", "lint", "test"],
     "backend": "json",
 }
 
@@ -13,12 +13,12 @@ HERE = Path(__file__).parent
 
 def task_format():
     """Reformat all files using black."""
-    return {"actions": [["black", HERE]], "verbosity": 1}
+    return {"actions": [["black", HERE], ["isort", HERE]], "verbosity": 1}
 
 
 def task_format_check():
     """Check, but not change, formatting using black."""
-    return {"actions": [["black", HERE, "--check"]], "verbosity": 1}
+    return {"actions": [["black", HERE, "--check"], ["isort", HERE, "--check-only"]], "verbosity": 1}
 
 
 def task_test():
