### 2021-10-08 5975b16
updated documentation to include example gallery automatically generated from Example Notebooks
