+++ docs/conf.py
@@ -27,6 +27,7 @@ __location__ = os.path.join(os.getcwd(), os.path.dirname(inspect.getfile(inspect
 from datetime import datetime
 from pathlib import Path
 
+sys.path.insert(0, os.path.abspath("."))
 sys.path.insert(0, os.path.join(__location__, "../src/biopsykit"))
 
 # -- Run sphinx-apidoc -------------------------------------------------------
@@ -83,7 +84,10 @@ extensions = [
     "sphinx.ext.doctest",
     "sphinx.ext.ifconfig",
     "sphinx.ext.mathjax",
-    # "sphinx_gallery.gen_gallery"
+    "sphinx_copybutton",
+    "sphinx_gallery.load_style",
+    "button",
+    "nbsphinx",
     # "sphinx_rtd_theme",
 ]
 
@@ -195,30 +199,31 @@ copyright = (
 
 # -- Copy README, Changelog, etc. --------------------------------------
 HERE = Path(__file__).parent
-with (HERE.parent.joinpath("README.md")).open() as f:
-    out = f.read()
-with (HERE.joinpath("README.md")).open("w+") as f:
-    f.write(out)
+EXAMPLE_NOTEBOOKS_DIR = HERE.joinpath("examples/_notebooks")
 
-with (HERE.parent.joinpath("coverage-badge.svg")).open() as f:
-    out = f.read()
-with (HERE.joinpath("coverage-badge.svg")).open("w+") as f:
-    f.write(out)
+shutil.copy(HERE.parent.joinpath("README.md"), HERE.joinpath("README.md"))
+shutil.copy(HERE.parent.joinpath("CHANGELOG.md"), HERE.joinpath("CHANGELOG.md"))
+shutil.copy(HERE.parent.joinpath("AUTHORS.md"), HERE.joinpath("source/about/AUTHORS.md"))
+shutil.copy(HERE.parent.joinpath("CONTRIBUTING.md"), HERE.joinpath("source/contributing/CONTRIBUTING.md"))
+shutil.copy(HERE.parent.joinpath("coverage-badge.svg"), HERE.joinpath("coverage-badge.svg"))
 
-with (HERE.parent.joinpath("CHANGELOG.md")).open() as f:
-    out = f.read()
-with (HERE.joinpath("CHANGELOG.md")).open("w+") as f:
-    f.write(out)
 
-with (HERE.parent.joinpath("AUTHORS.md")).open() as f:
-    out = f.read()
-with (HERE.joinpath("source/about/AUTHORS.md")).open("w+") as f:
-    f.write(out)
+def all_but_ipynb(dir, contents):
+    result = []
+    for c in contents:
+        if os.path.isfile(os.path.join(dir, c)) and (not c.endswith(".ipynb")):
+            result += [c]
+    return result
 
-with (HERE.parent.joinpath("CONTRIBUTING.md")).open() as f:
-    out = f.read()
-with (HERE.joinpath("source/contributing/CONTRIBUTING.md")).open("w+") as f:
-    f.write(out)
+
+shutil.rmtree(EXAMPLE_NOTEBOOKS_DIR, ignore_errors=True)
+shutil.copytree(HERE.parent.joinpath("examples"), EXAMPLE_NOTEBOOKS_DIR, ignore=all_but_ipynb)
+for file in EXAMPLE_NOTEBOOKS_DIR.glob("*.ipynb"):
+    with file.open() as f:
+        out = f.read()
+    out = out.replace("%matplotlib widget", "%matplotlib inline")
+    with file.open("w+") as f:
+        f.write(out)
 
 
 # The language for content autogenerated by Sphinx. Refer to documentation
@@ -299,6 +304,12 @@ version = ".".join(release.split(".")[:3])
 # so a file named "default.css" will overwrite the builtin "default.css".
[…]
