+++ docs/index.rst
@@ -4,7 +4,7 @@ BioPsyKit Documentation Overview
 Welcome to the documentation of **BioPsyKit** – your Python package for the analysis of biopsychological data.
 
 This documentation will help you to get up to speed quickly. Take a look at the :ref:`User Guide<user-guide>`, and the
-:ref:`Examples<examples>` sections for a guide on how to use **BioPsyKit**.
+:ref:`Example Gallery<examples>` sections for a guide on how to use **BioPsyKit**.
 
 .. toctree::
    :maxdepth: 1
