+++ docs/examples/index.rst
@@ -1,6 +1,108 @@
 .. _examples:
 
-Examples
-========
+===============
+Example Gallery
+===============
 
-coming soon...
+
+Physiological Signals
+=====================
+These examples demonstrate how to process and analyze different types of physiological data.
+
+.. nbgallery::
+    :caption: Physiological Signals
+    :name: physiological-signals
+    :glob:
+
+    _notebooks/ECG_Processing_Example
+    _notebooks/ECG_Analysis_Example
+    _notebooks/EEG_Example
+    _notebooks/CFT_Example
+
+
+Sleep Data
+==========
+These examples demonstrate how to process sleep data, either collected using inertial measurement units (IMUs) or
+using sleep trackers.
+
+.. nbgallery::
+    :caption: Sleep Data
+    :name: sleep-data
+    :glob:
+
+    _notebooks/Sleep_IMU_Example
+    _notebooks/Sleep_Analyzer_Example
+
+
+Questionnaire Data
+==================
+This example demonstrates how to process questionnaire data.
+
+.. nbgallery::
+    :caption: Questionnaire Data
+    :name: questionnaire-data
+    :glob:
+
+    _notebooks/Questionnaire_Example
+
+
+Saliva Data
+===========
+This example demonstrates how to process saliva data.
+
+.. nbgallery::
+    :caption: Saliva Data
+    :name: saliva-data
+    :glob:
+
+    _notebooks/Saliva_Example
+
+
+Psychological Protocols
+=======================
+These examples demonstrate how to generally handle data collected during psychological protocols (or other studies)
+and how to use the ``Protocol`` API in particular.
+
+.. nbgallery::
+    :caption: Psychological Protocols
+    :name: psychological-protocols
+    :glob:
+
+    _notebooks/ECG_Analysis_Example
+    _notebooks/Protocol_Example
+    _notebooks/CFT_Example
+
+
+Evaluation
+==========
+These examples demonstrate how to evaluate results either with statistical methods or with machine learning methods.
+
+.. nbgallery::
+    :caption: Evaluation
+    :name: evaluation-examples
+    :glob:
+
+    _notebooks/StatsPipeline_Plotting_Example
+    _notebooks/SklearnPipelinePermuter_Example
+
+
+Color Palettes
+==============
+This example shows how to generate color palettes used for plotting.
+
+.. nbgallery::
+    :caption: Color Palettes
+    :name: color-palettes
+    :glob:
+
+    _notebooks/Color_Example
+
+
+Miscellaneous
+=============
+.. nbgallery::
+    :caption: Miscellaneous
+    :name: miscellaneous-examples
+    :glob:
+
+    _notebooks/Log_Data_Example
