### 2021-10-08 0bbfb26
updated dev dependencies (needed to build documentation)
