+++ docs/requirements.txt
@@ -1,32 +1,6 @@
 # Requirements file for ReadTheDocs, check .readthedocs.yml.
-# To build the module reference correctly, make sure every external package
-# under `install_requires` in `setup.cfg` is also listed here!
-# sphinx_rtd_theme
-ipympl >= 0.5.6
-IPython >= 7.13.0
-ipywidgets >= 7.6.3
-matplotlib >= 3.3.0
-mne >= 0.21.0
-neurokit2 >= 0.0.40
-nilspodlib >= 3.1
-numpy >= 1
-pandas >= 1
-pingouin >= 0.3.10
-pytz >= 2020
-scipy >= 1, <1.6
 seaborn >= 0.11.1
-setuptools >= 54.2.0
-statannot >= 0.2.3
-tqdm >= 4.55
-typing_extensions >= 3.7
-black >= 19.10b0
 Sphinx >= 3.0.1
-pytest >= 6.2.1
 recommonmark >= 0.7.1
 sphinx>=3.2.1
 pydata-sphinx-theme>=0.4.1
-pytest-cov >= 2.8.1
-prospector >=1.2.0
-coverage >= 4.5.4
-doit >= 0.32.0
-mypy >= 0.790
