### 2021-10-08 0d88c7a
updated doc requirements
