+++ docs/requirements.txt
@@ -4,3 +4,6 @@ Sphinx >= 3.0.1
 recommonmark >= 0.7.1
 sphinx>=3.2.1
 pydata-sphinx-theme>=0.4.1
+sphinx-copybutton>=0.4.0
+nbsphinx>=0.8.7
+sphinx-gallery>=0.10.0
