+++ pyproject.toml
@@ -56,7 +56,7 @@ pytest-xdist = "^2.2.0"
 # Doc dependencies. They also need to be updated in ./docs/requirements.txt
 sphinx = "^3.2.1"
 numpydoc = "^1"
-sphinx-gallery = "^0.8.2"
+sphinx-gallery = "^0.10.0"
 sphinx_issues = "^1.2.0"
 pydata-sphinx-theme = "^0.6.2"
 recommonmark = "^0.6.0"
