+++ tests/test_example_data/test_example_data.py
@@ -9,16 +9,16 @@ import biopsykit.example_data
 from biopsykit.example_data import *
 from biopsykit.questionnaires import pss
 from biopsykit.questionnaires.utils import find_cols
-from biopsykit.utils._datatype_validation_helper import _assert_is_dtype, _assert_has_columns, _assert_has_index_levels
+from biopsykit.utils._datatype_validation_helper import _assert_has_columns, _assert_has_index_levels, _assert_is_dtype
 from biopsykit.utils.datatype_helper import (
-    is_subject_condition_dataframe,
-    is_saliva_raw_dataframe,
-    is_saliva_mean_se_dataframe,
-    is_hr_phase_dict,
-    is_sleep_endpoint_dataframe,
-    is_imu_dataframe,
     is_ecg_raw_dataframe,
+    is_hr_phase_dict,
     is_hr_subject_data_dict,
+    is_imu_dataframe,
+    is_saliva_mean_se_dataframe,
+    is_saliva_raw_dataframe,
+    is_sleep_endpoint_dataframe,
+    is_subject_condition_dataframe,
 )
 from biopsykit.utils.exceptions import ValueRangeError
 
