+++ src/biopsykit/plotting/__init__.py
@@ -1,5 +1,5 @@
 """Module providing various customized plotting functions."""
-from biopsykit.plotting.plotting import lineplot, stacked_barchart, feature_boxplot, multi_feature_boxplot
+from biopsykit.plotting.plotting import feature_boxplot, lineplot, multi_feature_boxplot, stacked_barchart
 
 __all__ = [
     "lineplot",
