+++ src/biopsykit/sleep/sleep_wake_detection/algorithms/cole_kripke_alternative.py
@@ -3,10 +3,10 @@ import numpy as np
 import pandas as pd
 
 from biopsykit.sleep.sleep_wake_detection.algorithms._base import _SleepWakeBase
+from biopsykit.sleep.sleep_wake_detection.utils import rescore
 from biopsykit.utils._types import arr_t
 from biopsykit.utils.array_handling import sanitize_input_1d
-from biopsykit.sleep.sleep_wake_detection.utils import rescore
-from biopsykit.utils.datatype_helper import _SleepWakeDataFrame, SleepWakeDataFrame
+from biopsykit.utils.datatype_helper import SleepWakeDataFrame, _SleepWakeDataFrame
 
 
 class ColeKripkeAlternative(_SleepWakeBase):
