+++ src/biopsykit/metadata/metadata.py
@@ -7,9 +7,10 @@ If ``data`` is a dataframe that contains more than the required two columns, e.g
 is passed, the required columns can be sliced by specifying them in the ``columns`` parameter.
 
 """
-from typing import Optional, Union, Sequence
+from typing import Optional, Sequence, Union
 
 import pandas as pd
+
 from biopsykit.utils._datatype_validation_helper import _assert_value_range
 
 __all__ = ["bmi", "whr"]
