### 2021-10-08 5d2258b
updated doc requirements
