+++ docs/requirements.txt
@@ -1,4 +1,5 @@
 # Requirements file for ReadTheDocs, check .readthedocs.yml.
+ipykernel>=5.2.0
 seaborn >= 0.11.1
 Sphinx >= 3.0.1
 recommonmark >= 0.7.1
