+++ src/biopsykit/questionnaires/utils.py
@@ -8,7 +8,6 @@ import numpy as np
 import pandas as pd
 from typing_extensions import Literal
 
-from biopsykit.questionnaires import questionnaires
 from biopsykit.utils._datatype_validation_helper import _assert_is_dtype, _assert_len_list, _assert_value_range
 from biopsykit.utils.dataframe_handling import wide_to_long as wide_to_long_utils
 
@@ -625,6 +624,8 @@ def compute_scores(
     >>> compute_scores(data, quest_dict)
 
     """
+    from biopsykit.questionnaires import questionnaires  # pylint:disable=import-outside-toplevel
+
     _assert_is_dtype(data, pd.DataFrame)
 
     df_scores = pd.DataFrame(index=data.index)
@@ -672,6 +673,8 @@ def get_supported_questionnaires() -> Dict[str, str]:
         dictionary with questionnaire names (keys) and description (values)
 
     """
+    from biopsykit.questionnaires import questionnaires  # pylint:disable=import-outside-toplevel
+
     funcs = dict(getmembers(questionnaires, isfunction))
     quests = {}
     for key, value in funcs.items():
