### 2021-10-08 844cb97
updated dependencies for docs
