+++ docs/requirements.txt
@@ -1,7 +1,9 @@
 # Requirements file for ReadTheDocs, check .readthedocs.yml.
+IPython>=7.13.0
 ipykernel>=5.2.0
+ipywidgets>=7.6.3
+ipympl>=0.7.0
 seaborn >= 0.11.1
-Sphinx >= 3.0.1
 recommonmark >= 0.7.1
 sphinx>=3.2.1
 pydata-sphinx-theme>=0.4.1
