### 2021-10-09 e340887
updated docs dependencies
