+++ docs/requirements.txt
@@ -10,3 +10,4 @@ pydata-sphinx-theme>=0.4.1
 sphinx-copybutton>=0.4.0
 nbsphinx>=0.8.7
 sphinx-gallery>=0.10.0
+mne>=0.23.0
