### 2021-10-09 a03aca8
added ipython kernel specs for doc notebooks
