### 2021-10-09 b038938
fixed linting and type checking
