+++ .github/workflows/test-and-lint.yml
@@ -32,20 +32,8 @@ jobs:
     - name: Testing
       run: |
         poetry run doit test
-    - name: Create Coverage Badge
-      if: ${{ matrix.python-version == 3.8 }}
-      run: |
-        poetry run coverage-badge -o coverage-badge.svg -f
-    - name: Commit Coverage Badge
-      uses: stefanzweifel/git-auto-commit-action@v4
-      if: ${{ github.event_name == 'push' }} && ${{ matrix.python-version == 3.8 }}
+    - name: "Upload coverage to Codecov"
+      uses: codecov/codecov-action@v2
       with:
-        commit_message: added coverage badge
-        commit_user_name: GitHub Actions
-        commit_user_email: github-actions@users.noreply.github.com
-    - name: Push changes
-      uses: ad-m/github-push-action@master
-      if: ${{ matrix.python-version == 3.8 }}
-      with:
-        github_token: ${{ secrets.GITHUB_TOKEN }}
-        branch: ${{ github.ref }}
+        fail_ci_if_error: true
+
