### 2021-10-09 9a193fd
Docs for RTD (#17)
* updated doc requirements

* updated doc requirements

* added default kernel name

* updated dependencies for docs

* notebook test

* notebook test

* updated docs dependencies

* changed notebook for docs

* changed notebook for docs

* added ipython kernel specs for doc notebooks

* updated example notebooks

* updated font-family style in example notebooks
