+++ docs/requirements.txt
@@ -1,6 +1,13 @@
 # Requirements file for ReadTheDocs, check .readthedocs.yml.
+IPython>=7.13.0
+ipykernel>=5.2.0
+ipywidgets>=7.6.3
+ipympl>=0.7.0
 seaborn >= 0.11.1
-Sphinx >= 3.0.1
 recommonmark >= 0.7.1
 sphinx>=3.2.1
 pydata-sphinx-theme>=0.4.1
+sphinx-copybutton>=0.4.0
+nbsphinx>=0.8.7
+sphinx-gallery>=0.10.0
+mne>=0.23.0
