### 2021-10-10 541a34c
updated example notebooks
