### 2021-10-10 f97f3ce
updated example notebooks
