### 2021-10-11 mad-lab-fau/BioPsyKit#14 @richrobe
Hi @espenhgn , yes, that's the expected output because the ECG sequences are too short to compute certain HRV parameter. The same warning is also produced for the example in the [documentation](https://biopsykit.readthedocs.io/en/latest/examples/_notebooks/ECG_Analysis_Example.html).
