### 2021-10-11 c5e36ce
added pandoc as dev and doc dependency
