+++ docs/requirements.txt
@@ -9,5 +9,6 @@ sphinx>=3.2.1
 pydata-sphinx-theme>=0.4.1
 sphinx-copybutton>=0.4.0
 nbsphinx>=0.8.7
+pandoc>=1.1.0
 sphinx-gallery>=0.10.0
 mne>=0.23.0
