### 2021-10-11 ea697f2
added further information on doit update_version task to update the package version
