+++ dodo.py
@@ -83,11 +83,15 @@ def update_version(version):
         .strip()
         .split(" ", 1)[1]
     )
-    update_version_strings(HERE / "src/biopsykit/__init__.py", new_version)
+    update_version_strings(HERE.joinpath("src/biopsykit/__init__.py"), new_version)
 
 
 def task_update_version():
-    """Bump the version in pyproject.toml and biopsykit.__init__ ."""
+    """Bump the version in pyproject.toml and biopsykit.__init__ .
+
+    Pass the options "major", "minor", or "patch" with the `-v` argument to bump the major, minor, or patch version,
+    respectively.
+    """
     return {
         "actions": [(update_version,)],
         "params": [{"name": "version", "short": "v", "default": None}],
