### 2021-10-11 46bbc9d
added note to manually install pandoc before building the documentation locally
