+++ paper/paper.md
@@ -35,41 +35,37 @@ bibliography: paper.bib
 ---
 
 # Summary
-_Biopsychology_ is a field of psychology that analyzes how biological processes interact with behaviour, emotion, 
-cognition, and other mental processes. Biopsychology covers, among others, the topics of sensation and perception, 
-emotion regulation, movement (and control of such), sleep and biological rhythms, as well as acute and chronic stress.
+_Biopsychology_ is a field of psychology that analyzes how biological processes interact with behaviour, emotion, cognition, and other mental processes. 
+Biopsychology covers, among others, the topics of sensation and perception, emotion regulation, movement (and control of such), sleep and biological rhythms, as well as acute and chronic stress.
 
-To assess the interaction between biological and mental processes a variety of different modalities are used in the 
-field of biopsychology, such as _electrophysiology_, assessed, for instance, via electrocardiography (ECG), 
-electrodermal activity (EDA), or electroencephalography (EEG), _sleep, activity and movement_, assessed via inertial measurement units (IMUs), _neuroendocrine and inflammatory 
-biomarker_, assessed by saliva and blood samples, as well as _self-reports_, assessed via psychological questionnaires. 
+To assess the interaction between biological and mental processes a variety of different modalities are used in the field of biopsychology, such as _electrophysiology_, assessed, for instance, via electrocardiography (ECG), electrodermal activity (EDA), or electroencephalography (EEG), _sleep, activity and movement_, assessed via inertial measurement units (IMUs), _neuroendocrine and inflammatory biomarkers_, assessed by saliva and blood samples, as well as _self-reports_, assessed via psychological questionnaires. 
 
 These different modalities are collected either "in the lab", during standardized laboratory protocols, or "in the wild", during unsupervised protocols in home environments. 
 The collected data are typically analyzed using statistical methods, or, more recently, using machine learning methods.
 
 
-While some software packages exist that allow for the analysis of single data modalities, such as electrophysiological data, or sleep, activity and movement data, no packages are available for the analysis of other modalities, such as neuroendocrine and inflammatory biomarker, and self-reports. In order to fill this gap, and, simultaneously, to combine all required tools analyzing biopsychological data from beginning to end into one single Python package, we developed `BioPsyKit`.
+While some software packages exist that allow for the analysis of single data modalities, such as electrophysiological data, or sleep, activity and movement data, no packages are available for the analysis of other modalities, such as neuroendocrine and inflammatory biomarker, and self-reports. 
+In order to fill this gap, and, simultaneously, to combine all required tools analyzing biopsychological data from beginning to end into one single Python package, we developed `BioPsyKit`.
 
 
-# Statement of need
-Researchers in biopsychology often face the challenge to analyze data from different assessment modalities during experiments in order to capture the 
-complex interaction between biological and mental processes. 
+# Statement of Need
+Researchers in biopsychology often face the challenge of analyzing data from different assessment modalities during experiments in order to capture the complex interaction between biological and mental processes. 
 
-One example might be collecting (electro)physiological (e.g., ECG) or 
-salivary biomarkers (e.g., cortisol) during an acute stress protocol, and investigating the correlation 
-between biomarkers and psychometric data assessed via self-reports, such as perceived stress, state anxiety, 
-or positive/negative affect. Another example is the assessment of relationships between sleep and neuroendocrine responses in the morning. To assess the 
[…]
