### 2021-10-12 2ec9aab
added citation instructions to readme
