### 2021-10-12 91d1132
0.3.2: Final version for JOSS submission
