+++ src/biopsykit/__init__.py
@@ -40,7 +40,7 @@ __all__ = [
     "utils",
 ]
 
-__version__ = "0.3.1"
+__version__ = "0.3.2"
 
 
 def version() -> None:
