### 2021-10-12 08ba524
added correct JOSS citation to README
