+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.3.2 – October 12 2021
+- final version of BioPsyKit for submission to the [Journal of Open Source Software](https://joss.theoj.org/).
+- improved documentation (e.g., included Example Gallery)
+
 ## Version 0.3.1 – October 5 2021
 - `biopsykit.utils.file_handling.get_subject_dirs()`: replaced `glob`-style matching of regex strings 
   by `re`-style matching. This allows more flexibility when filtering for directories that match a specific pattern. 
