### 2021-10-12 dd356e2
fixed correct parsing of sleep analyzer data during import (because sleep analyzer data can have duplicate index values, which will then be exploded wrongly)
